Move RadioOption to viewmodels package
