+++ src/main/g8/app/views/components/input_radio.scala.html
@@ -1,4 +1,4 @@
-@import utils.RadioOption
+@import viewmodels.RadioOption
 
 @(
         field: Field,
