+++ src/main/g8/test/viewmodels/RadioOptionSpec.scala
@@ -1,15 +1,18 @@
-package utils
+package viewmodels
 
 import base.SpecBase
 
 class RadioOptionSpec extends SpecBase {
 
   "Radio Option" must {
+
     "build correctly from a key prefix and option" in {
+
       val radioOption = RadioOption("prefix", "option")
-      radioOption.id mustBe "prefix.option"
-      radioOption.value mustBe "option"
-      radioOption.messageKey mustBe "prefix.option"
+
+      radioOption.id mustEqual "prefix.option"
+      radioOption.value mustEqual "option"
+      radioOption.messageKey mustEqual "prefix.option"
     }
   }
 }
